+++ db-service/test/cqn4sql/expand.test.js
@@ -278,6 +278,7 @@ describe('Unfold expands on associations to special subselects', () => {
   })
 
   it('nested expand with unmanaged backlink', () => {
+    const localized_ = cds.unfold ? '' : 'localized.'
     let expandQuery = CQL`select from bookshop.DataRestrictions {
       *,
       dataRestrictionAccessGroups {
@@ -288,11 +289,11 @@ describe('Unfold expands on associations to special subselects', () => {
         }
       }
     }`
-    let expected = CQL`
-      select from localized.bookshop.DataRestrictions as DataRestrictions {
+    let expected = CQL(`
+      select from ${localized_}bookshop.DataRestrictions as DataRestrictions {
         DataRestrictions.ID,
         (
-          select from localized.bookshop.DataRestrictionAccessGroups as dataRestrictionAccessGroups {
+          select from ${localized_}bookshop.DataRestrictionAccessGroups as dataRestrictionAccessGroups {
             dataRestrictionAccessGroups.dataRestrictionID,
             dataRestrictionAccessGroups.accessGroupID,
             (
@@ -303,10 +304,10 @@ describe('Unfold expands on associations to special subselects', () => {
           } where DataRestrictions.ID = dataRestrictionAccessGroups.dataRestrictionID
         ) as dataRestrictionAccessGroups
       }
-    `
+    `)
     // seems to only happen with the `for.nodejs(…)` compiled model
     expandQuery.SELECT.localized = true
-    expect(JSON.parse(JSON.stringify(cqn4sql(expandQuery, cds.compile.for.nodejs(model))))).to.deep.equal(expected)
+    expect(JSON.parse(JSON.stringify(cqn4sql(expandQuery, cds.compile.for.nodejs(JSON.parse(JSON.stringify(model))))))).to.deep.equal(expected)
   })
 
   it('add where exists <assoc> shortcut to expand subquery where condition', () => {
@@ -962,7 +963,7 @@ describe('Unfold expands on associations to special subselects', () => {
           author.name
         } as bookInfos
       }`
-      const qx = CQL`SELECT from bookshop.Books as Books 
+      const qx = CQL`SELECT from bookshop.Books as Books
         left join bookshop.Authors as author on author.ID = Books.author_ID
       {
         Books.ID,
@@ -1021,7 +1022,7 @@ describe('Unfold expands on associations to special subselects', () => {
         }
       }
     `
-    let transformed = cqn4sql(q, cds.compile.for.nodejs(model))
+    let transformed = cqn4sql(q, cds.compile.for.nodejs(JSON.parse(JSON.stringify(model))))
     expect(JSON.parse(JSON.stringify(transformed))).to.deep.eql(CQL`
       SELECT from Collaborations as Collaborations {
         Collaborations.id,
@@ -1121,7 +1122,7 @@ describe('expand on structure part II', () => {
         (SELECT department.id, department.name from Department as department where Employee.department_id = department.id) as department,
         (SELECT assets.id, assets.descr from Assets as assets where Employee.id = assets.owner_id) as assets
     }`
-    expect(JSON.parse(JSON.stringify(cqn4sql(expandQuery, cds.compile.for.nodejs(model))))).to.eql(expected)
+    expect(JSON.parse(JSON.stringify(cqn4sql(expandQuery, cds.compile.for.nodejs(JSON.parse(JSON.stringify(model))))))).to.eql(expected)
   })
 
   it('multi expand with star but foreign key does not survive in structured mode', () => {