+++ hana/lib/HANAService.js
@@ -98,7 +98,12 @@ class HANAService extends SQLService {
     this.ensureDBC().set(variables)
   }
 
-  async onSELECT({ query, data }) {
+  async onSELECT(req) {
+    const { query, data } = req
+    if (!query.target || query.target._unresolved) {
+      return super.onSELECT(req)
+    }
+
     // REVISIT: disable this for queries like (SELECT 1)
     // Will return multiple rows with objects inside
     query.SELECT.expand = 'root'
@@ -119,6 +124,7 @@ class HANAService extends SQLService {
   }
 
   async onINSERT({ query, data }) {
+    try {
     const { sql, entries, cqn } = this.cqn2sql(query, data)
     if (!sql) return // Do nothing when there is nothing to be done
     const ps = await this.prepare(sql)
@@ -129,6 +135,17 @@ class HANAService extends SQLService {
         : ps.run(entries[0])
       : ps.run())
     return new this.class.InsertResults(cqn, results)
+    } catch (err) {
+      throw _not_unique(err, 'ENTITY_ALREADY_EXISTS')
+    }
+  }
+
+  async onUPDATE(req) {
+    try {
+      return await super.onUPDATE(req)
+    } catch (err) {
+      throw _not_unique(err, 'UNIQUE_CONSTRAINT_VIOLATION') || err
+    }
   }
 
   // Allow for running complex expand queries in a single statement
@@ -249,7 +266,7 @@ class HANAService extends SQLService {
       this.temporary = this.temporary || []
       this.temporaryValues = this.temporaryValues || []
 
-      const { limit, one, orderBy, expand, columns, localized, count, parent } = q.SELECT
+      const { limit, one, orderBy, expand, columns = ['*'], localized, count, parent } = q.SELECT
 
       const walkAlias = q => {
         if (q.args) return q.as || walkAlias(q.args[0])
@@ -393,7 +410,9 @@ class HANAService extends SQLService {
                 expands[this.column_name(x)] = x.SELECT.one ? null : []
 
                 const parent = src
-                x.element._foreignKeys.forEach(k => {
+                let fkeys = x.element._foreignKeys
+                if (typeof fkeys === 'function') fkeys = fkeys.call(x.element)
+                fkeys.forEach(k => {
                   if (!parent.SELECT.columns.find(c => this.column_name(c) === k.parentElement.name)) {
                     parent.SELECT.columns.push({ ref: [parent.as, k.parentElement.name] })
                   }
@@ -508,9 +527,10 @@ class HANAService extends SQLService {
       const entity = q.target?.['@cds.persistence.name'] || this.name(q.target?.name || INSERT.into.ref[0])
 
       const elements = q.elements || q.target?.elements
-      if (!elements && !INSERT.entries?.length) {
-        return // REVISIT: mtx sends an insert statement without entries and no reference entity
+      if (!elements) {
+        return super.INSERT_entries(q)
       }
+
       const columns = elements
         ? ObjectKeys(elements).filter(c => c in elements && !elements[c].virtual && !elements[c].value && !elements[c].isAssociation)
         : ObjectKeys(INSERT.entries[0])
@@ -569,6 +589,10 @@ class HANAService extends SQLService {
       // The problem with Simple INSERT is the type mismatch from csv files
       // Recommendation is to always use entries
       const elements = q.elements || q.target?.elements
+      if (!elements) {
+        return super.INSERT_rows(q)
+      }
+
       const columns = INSERT.columns || (elements && ObjectKeys(elements))
       const entries = new Array(INSERT.rows.length)
       const rows = INSERT.rows
@@ -585,13 +609,17 @@ class HANAService extends SQLService {
     }
 
     UPSERT(q) {
-      let { UPSERT } = q,
-        sql = this.INSERT({ __proto__: q, INSERT: UPSERT })
+      const { UPSERT } = q
+      const sql = this.INSERT({ __proto__: q, INSERT: UPSERT })
+
+      // If no definition is available fallback to INSERT statement
+      const elements = q.elements || q.target?.elements
+      if (!elements) {
+        return (this.sql = sql)
+      }
 
       // REVISIT: should @cds.persistence.name be considered ?
       const entity = q.target?.['@cds.persistence.name'] || this.name(q.target?.name || INSERT.into.ref[0])
-      const elements = q.elements || q.target?.elements
-
       const dataSelect = sql.substring(sql.indexOf('WITH'))
 
       // Calculate @cds.on.insert
@@ -605,7 +633,7 @@ class HANAService extends SQLService {
       const keyCompare =
         keys &&
         Object.keys(keys)
-          .filter(k => !keys[k].isAssociation)
+          .filter(k => !keys[k].isAssociation && !keys[k].virtual)
           .map(k => `NEW.${this.quote(k)}=OLD.${this.quote(k)}`)
           .join(' AND ')
 
@@ -636,19 +664,24 @@ class HANAService extends SQLService {
     }
 
     where(xpr) {
-      return this.xpr({ xpr }, ' = TRUE')
+      xpr = { xpr }
+      const suffix = this.is_comparator(xpr) ? '' : ' = TRUE'
+      return `${this.xpr(xpr)}${suffix}`
     }
 
     having(xpr) {
-      return this.xpr({ xpr }, ' = TRUE')
+      return this.where(xpr)
     }
 
-    xpr({ xpr, _internal }, caseSuffix = '') {
+    xpr(_xpr, caseSuffix = '') {
+      const { xpr, _internal } = _xpr
       // Maps the compare operators to what to return when both sides are null
       const compareOperators = {
-        '=': true,
+        '==': true,
         '!=': false,
         // These operators are not allowed in column expressions
+        /* REVISIT: Only adjust these operators when inside the column expression
+        '=': null,
         '>': null,
         '<': null,
         '<>': null,
@@ -656,26 +689,29 @@ class HANAService extends SQLService {
         '<=': null,
         '!<': null,
         '!>': null,
+        */
       }
 
+      let endWithCompare = false
       if (!_internal) {
         for (let i = 0; i < xpr.length; i++) {
-          const x = xpr[i]
+          let x = xpr[i]
           if (typeof x === 'string') {
+            // Convert =, == and != into is (not) null operator where required
+            x = xpr[i] = super.operator(xpr[i], i, xpr)
+
             // HANA does not support comparators in all clauses (e.g. SELECT 1>0 FROM DUMMY)
             // HANA does not have an 'IS' or 'IS NOT' operator
             if (x in compareOperators) {
+              endWithCompare = true
               const left = xpr[i - 1]
               const right = xpr[i + 1]
               const ifNull = compareOperators[x]
 
-              const compare = {
-                xpr: [left, x, right],
-                _internal: true,
-              }
+              const compare = [left, x, right]
 
               const expression = {
-                xpr: ['CASE', 'WHEN', compare, 'Then', { val: true }, 'WHEN', 'NOT', compare, 'Then', { val: false }],
+                xpr: ['CASE', 'WHEN', ...compare, 'THEN', { val: true }, 'WHEN', 'NOT', ...compare, 'THEN', { val: false }],
                 _internal: true,
               }
 
@@ -686,11 +722,8 @@ class HANAService extends SQLService {
                   xpr: [
                     'CASE',
                     'WHEN',
-                    {
-                      xpr: [left, 'IS', 'NULL', 'AND', right, 'IS', 'NULL'],
-                      _internal: true,
-                    },
-                    'Then',
+                    ...[left, 'IS', 'NULL', 'AND', right, 'IS', 'NULL'],
+                    'THEN',
                     { val: ifNull },
                     'ELSE',
                     { val: !ifNull },
@@ -703,7 +736,7 @@ class HANAService extends SQLService {
 
               xpr[i - 1] = ''
               xpr[i] = expression
-              xpr[i + 1] = caseSuffix || ''
+              xpr[i + 1] = ''
             }
           }
         }
@@ -713,39 +746,83 @@ class HANAService extends SQLService {
       for (let i = 0; i < xpr.length; ++i) {
         const x = xpr[i]
         if (typeof x === 'string') {
+          const up = x.toUpperCase()
+          if (up in logicOperators) {
+            // Force current expression to end with a comparison
+            endWithCompare = true
+          }
+          if (endWithCompare && (up in caseOperators || up === ')')) {
+            endWithCompare = false
+          }
           sql.push(this.operator(x, i, xpr))
         } else if (x.xpr) sql.push(`(${this.xpr(x, caseSuffix)})`)
         // default
         else sql.push(this.expr(x))
       }
 
-      // HANA does not allow WHERE TRUE so when the expression is only a single entry "= TRUE" is appended
-      if (caseSuffix && xpr.length === 1) {
-        sql.push(caseSuffix)
+      if (endWithCompare) {
+        const suffix = this.operator('OR', xpr.length, xpr).slice(0, -3)
+        if (suffix) {
+          sql.push(suffix)
+        }
       }
 
       return `${sql.join(' ')}`
     }
 
     operator(x, i, xpr) {
+      const up = x.toUpperCase()
       // Add "= TRUE" before THEN in case statements
-      // As all valid comparators are converted to booleans as SQL specifies
-      if (x in { THEN: 1, then: 1 }) return ` = TRUE ${x}`
-      if ((x in { LIKE: 1, like: 1 } && is_regexp(xpr[i + 1]?.val)) || x === 'regexp') return 'LIKE_REGEXPR'
+      if (
+        up in logicOperators &&
+        !this.is_comparator({ xpr }, i - 1)
+      ) {
+        return ` = TRUE ${x}`
+      }
+      if (
+        (up === 'LIKE' && is_regexp(xpr[i + 1]?.val)) ||
+        up === 'REGEXP'
+      ) return 'LIKE_REGEXPR'
       else return x
     }
 
-    list(list) {
-      const first = list.list[0]
-      // If the list only contains of lists it is replaced with a json function and a placeholder
-      if (this.values && first.list && !first.list.find(v => !v.val)) {
-        const extraction = first.list.map((v, i) => `"${i}" ${this.constructor.InsertTypeMap[typeof v.val]()} PATH '$.${i}'`)
-        this.values.push(JSON.stringify(list.list.map(l => l.list.reduce((l, c, i) => { l[i] = c.val; return l }, {}))))
-        return `(SELECT * FROM JSON_TABLE(?, '$' COLUMNS(${extraction})))`
+    get is_distinct_from_() { return '!=' }
+    get is_not_distinct_from_() { return '==' }
+
+    /**
+     * Checks if the xpr is a comparison or a value
+     * @param {} xpr
+     * @returns
+     */
+    is_comparator({ xpr }, start) {
+      const local = start != null
+      for (let i = start ?? xpr.length; i > -1; i--) {
+        const cur = xpr[i]
+        if (cur == null) continue
+        if (typeof cur === 'string') {
+          const up = cur.toUpperCase()
+          // When a compare operator is found the expression is a comparison
+          if (up in compareOperators || (!local && up in logicOperators)) return true
+          // When a case operator is found it is the start of the expression
+          if (up in caseOperators) break
+          continue
+        }
+        if ('xpr' in cur) return this.is_comparator(cur)
       }
-      // Call super for normal SQL behavior
-      return super.list(list)
+      return false
     }
+
+        list(list) {
+          const first = list.list[0]
+          // If the list only contains of lists it is replaced with a json function and a placeholder
+          if (this.values && first.list && !first.list.find(v => !v.val)) {
+            const extraction = first.list.map((v, i) => `"${i}" ${this.constructor.InsertTypeMap[typeof v.val]()} PATH '$.V${i}'`)
+            this.values.push(JSON.stringify(list.list.map(l => l.list.reduce((l, c, i) => { l[`V${i}`] = c.val; return l }, {}))))
+            return `(SELECT * FROM JSON_TABLE(?, '$' COLUMNS(${extraction})))`
+          }
+          // Call super for normal SQL behavior
+          return super.list(list)
+        }
 
     quote(s) {
       // REVISIT: casing in quotes when reading from entities it uppercase
@@ -775,10 +852,13 @@ class HANAService extends SQLService {
       const requiredColumns = !elements
         ? []
         : Object.keys(elements)
-          .filter(
-            e =>
-              (elements[e]?.[annotation] || (!isUpdate && elements[e]?.default)) && !columns.find(c => c.name === e),
-          )
+          .filter(e => {
+            if (elements[e]?.virtual) return false
+            if (columns.find(c => c.name === e)) return false
+            if (elements[e]?.[annotation]) return true
+            if (!isUpdate && elements[e]?.default) return true
+            return false
+          })
           .map(name => ({ name, sql: 'NULL' }))
 
       const keyZero = this.quote(
@@ -795,8 +875,7 @@ class HANAService extends SQLService {
         const val = _managed[element[annotation]?.['=']]
         let managed
         if (val) managed = this.func({ func: 'session_context', args: [{ val, param: false }] })
-        const type = this.insertType4(element)
-        let extract = sql ?? `${this.quote(name)} ${type} PATH '$.${name}'`
+        let extract = sql ?? `${this.quote(name)} ${this.insertType4(element)} PATH '$.${name}'`
         if (!isUpdate) {
           const d = element.default
           if (d && (d.val !== undefined || d.ref?.[0] === '$now')) {
@@ -926,7 +1005,10 @@ class HANAService extends SQLService {
     return super.onPlainSQL(req, next)
   }
 
-  onBEGIN() { }
+  onBEGIN() {
+    DEBUG?.('BEGIN')
+    return this.dbc?.begin()
+  }
 
   onCOMMIT() {
     DEBUG?.('COMMIT')
@@ -959,7 +1041,7 @@ class HANAService extends SQLService {
 
       const stmt = await this.dbc.prepare(createContainerDatabase)
       const res = await stmt.run([creds.user, creds.password, creds.containerGroup, !clean])
-      DEBUG?.(res.map(r => r.MESSAGE).join('\n'))
+      res && DEBUG?.(res.changes.map(r => r.MESSAGE).join('\n'))
     } finally {
       if (this.dbc) {
         // Release table lock
@@ -1001,7 +1083,7 @@ class HANAService extends SQLService {
 
       const stmt = await this.dbc.prepare(createContainerTenant.replaceAll('{{{GROUP}}}', creds.containerGroup))
       const res = await stmt.run([creds.user, creds.password, creds.schema, !clean])
-      res && DEBUG?.(res.map(r => r.MESSAGE).join('\n'))
+      res && DEBUG?.(res.changes.map(r => r.MESSAGE).join('\n'))
     } finally {
       await this.dbc.disconnect()
       delete this.dbc
@@ -1020,12 +1102,52 @@ Buffer.prototype.toJSON = function () {
   return this.toString('hex')
 }
 
+function _not_unique(err, code) {
+  if (err.code === 301)
+    return Object.assign(err, {
+      originalMessage: err.message, // FIXME: required because of next line
+      message: code, // FIXME: misusing message as code
+      code: 400, // FIXME: misusing code as (http) status
+    })
+  return err
+}
+
 const is_regexp = x => x?.constructor?.name === 'RegExp' // NOTE: x instanceof RegExp doesn't work in repl
 const ObjectKeys = o => (o && [...ObjectKeys(o.__proto__), ...Object.keys(o)]) || []
 const _managed = {
   '$user.id': '$user.id',
   $user: '$user.id',
   $now: '$now',
+}
+
+const caseOperators = {
+  'CASE': 1,
+  'WHEN': 1,
+  'THEN': 1,
+  'ELSE': 1,
+}
+const logicOperators = {
+  'THEN': 1,
+  'AND': 1,
+  'OR': 1,
+}
+const compareOperators = {
+  '=': 1,
+  '==': 1,
+  '!=': 1,
+  '>': 1,
+  '<': 1,
+  '<>': 1,
+  '>=': 1,
+  '<=': 1,
+  '!<': 1,
+  '!>': 1,
+  'IS': 1,
+  'IN': 1,
+  'LIKE': 1,
+  'IS NOT': 1,
+  'EXISTS': 1,
+  'BETWEEN': 1,
 }
 
 module.exports = HANAService